GenAI-Projects/1-API-vs-Local-Models/requirements.txt
openai==1.30.1
anthropic==0.25.0
tokenizers==0.22.2
google-generativeai==0.3.0
transformers==4.57.6
huggingface_hub==0.36.2
ollama==0.1.7
python-dotenv==1.0.0